Moved prize pages to backoffice folder and style refactor
