+++ packages/nextjs/app/backoffice/organizations/_components/PrizeTable.tsx
